% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitivityAnalyses_DP.R
\name{computeMaxDifference_LS}
\alias{computeMaxDifference_LS}
\title{Sensitivity analysis for MTL with least-square loss and joint feature selection and low-rank structure, respectively}
\usage{
computeMaxDifference_LS(X, Y, W, C, funcVal, nSimulations)
}
\arguments{
\item{X}{A set of \eqn{T} \eqn{(N_t \times P)} feature matrices, \eqn{t \in \{1,\ldots,T\}}}

\item{Y}{A set of \eqn{T} response vectors \eqn{\in \mathbb{R}^{N_t}}, \eqn{t \in \{1,\ldots,T\}}}

\item{W}{A \eqn{(P \times T)} coefficient matrix}

\item{C}{The hyper-parameter associated with L2 term in the respective MTL model}

\item{funcVal}{The original function value}

\item{nSimulations}{number of runs in the sensitivity analysis involved in the differential privacy mechanism}
}
\value{
A list of two elements named maxDifference (maximum difference between an original objective function value and a modified objective function value obtained when removing one data point from the input data, across \code{nSimulations} runs) and originalValue (original objective function value), respectively
}
\description{
Sensitivity analysis for MTL with least-square loss and joint feature selection and low-rank structure, respectively: Compute the difference between an original objective function value and a modified objective function value obtained when removing one data point from the input data, perform this procedure for in total \code{nSimulations} data point removals and calculate the maximum of the differences over all \code{nSimulations} repetitions
}
\examples{
set.seed(24)
X0<-list(matrix(rnorm(100000),nrow=200,ncol=500),matrix(rnorm(150000),nrow=300,ncol=500))
set.seed(24)
Y0<-list(rnorm(200),rnorm(300))
set.seed(24)
W0<-matrix(rnorm(1000),nrow=500,ncol=2)
C0<-1
funcVal0<-0.375
n<-100
maxdiff<-computeMaxDifference_LS(X=X0, Y=Y0, W=W0, C=C0,funcVal=funcVal0, nSimulations=n)

}
